Add GitHub Action for updating libstdc++ headers
Introduced a new GitHub Action workflow to automate the process of updating libstdc++ headers. This workflow includes steps for installing necessary dependencies, fetching the gcc source, configuring and building the headers, and finally pushing the updates to the repository.

Updated the README with configuration details necessary for `features.h`.

Fixed git url to push

.

Rename workflow file for clarity

Renamed `update-headers.yml` to `ci.yml` to better reflect the file's purpose in the continuous integration pipeline. This improves readability and maintainability of the workflow configurations.

.

.